+++ src/module/datasworn2/import/ids.ts
@@ -2,15 +2,15 @@ import shajs from 'sha.js'
 
 // Import a local copy of the legacy ID maps
 export const LegacyToDataswornIds: Record<string, string> = {
-	...require('../../../../../datasworn/src/legacy_ids/dataforged/classic/assets.json'),
-	...require('../../../../../datasworn/src/legacy_ids/dataforged/classic/moves.json'),
-	...require('../../../../../datasworn/src/legacy_ids/dataforged/classic/npcs.json'),
-	...require('../../../../../datasworn/src/legacy_ids/dataforged/classic/oracles.json'),
-	...require('../../../../../datasworn/src/legacy_ids/dataforged/starforged/assets.json'),
-	...require('../../../../../datasworn/src/legacy_ids/dataforged/starforged/moves.json'),
-	...require('../../../../../datasworn/src/legacy_ids/dataforged/starforged/oracles.json'),
-	...require('../../../../../datasworn/src/legacy_ids/dataforged/starforged/npcs.json'),
-	...require('../../../../../datasworn/src/legacy_ids/dataforged/starforged/truths.json')
+	...require('./legacy_ids/classic/assets.json'),
+	...require('./legacy_ids/classic/moves.json'),
+	...require('./legacy_ids/classic/npcs.json'),
+	...require('./legacy_ids/classic/oracles.json'),
+	...require('./legacy_ids/starforged/assets.json'),
+	...require('./legacy_ids/starforged/moves.json'),
+	...require('./legacy_ids/starforged/oracles.json'),
+	...require('./legacy_ids/starforged/npcs.json'),
+	...require('./legacy_ids/starforged/truths.json')
 }
 export const DataswornToLegacyIds: Record<string, string> = Object.fromEntries(
 	Object.entries(LegacyToDataswornIds).map(([k, v]) => [v, k])